+++ src/main/java/com/code/tama/tts/server/tardis/control_lists/ControlEntityRecord.java
@@ -3,9 +3,9 @@ package com.code.tama.tts.server.tardis.control_lists;
 
 import net.minecraft.world.phys.AABB;
 
-public record ControlEntityRecord(float cx, float cy, float cz, // pivot = center of cube
-		float hw, float hh, float hd, // half-extents (width, height, depth)
-		float yawDeg, // rotation around Y axis in degrees
+public record ControlEntityRecord(double cx, double cy, double cz, // pivot = center of cube
+		double hw, double hh, double hd, // half-extents (width, height, depth)
+		double yawDeg, // rotation around Y axis in degrees
 		int ID) {
 	/** Reconstruct the local-space AABB (un-rotated, centered on origin). */
 	public AABB toLocalAABB() {
